requirements.txt
# Add project dependencies here, e.g.:
# flask==2.3.2
dbt-core==1.11.2
dbt-snowflake==1.11.0
